@@ -1,5 +1,5 @@
 {
-    "name": "silber/page-cache",
+    "name": "siteorigin/laravel-page-cache",
     "description": "Caches responses as static files on disk for lightning fast page loads.",
     "keywords": ["laravel", "cache"],
     "license": "MIT",
@@ -7,29 +7,33 @@
         {
             "name": "Joseph Silber",
             "email": "contact@josephsilber.com"
+        },
+        {
+            "name": "Greg Priday",
+            "email": "greg@siteorigin.com"
         }
     ],
     "autoload": {
         "psr-4": {
-            "Silber\\PageCache\\": "src/"
+            "SiteOrigin\\PageCache\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Silber\\PageCache\\Tests\\": "tests"
+            "SiteOrigin\\PageCache\\Tests\\": "tests"
         }
     },
     "require": {
-        "php": ">=5.5.9",
-        "illuminate/contracts": "5.0 - 5.8|^6.0|^7.0|^8.0",
-        "illuminate/filesystem": "5.0 - 5.8|^6.0|^7.0|^8.0",
-        "symfony/http-foundation": "2.6 - 4|^5.0"
+        "php": ">=7.4",
+        "illuminate/contracts": "^8.0",
+        "illuminate/filesystem": "^8.0",
+        "symfony/http-foundation": "^5.2",
+        "spatie/crawler": "^6.0"
     },
     "require-dev": {
-        "illuminate/container": "5.0 - 5.8|^6.0|^7.0|^8.0",
-        "mockery/mockery": "^0.9.5",
-        "phpunit/phpunit": "^4.8",
-        "symfony/var-dumper": "^4.4|^5.0"
+        "mockery/mockery": "^1.4",
+        "phpunit/phpunit": "^9.5",
+        "symfony/var-dumper": "^5.2"
     },
     "suggest": {
         "illuminate/console": "Allows clearing the cache via artisan"
@@ -37,7 +41,7 @@
     "extra": {
         "laravel": {
             "providers": [
-                "Silber\\PageCache\\LaravelServiceProvider"
+                "SiteOrigin\\PageCache\\PageCacheServiceProvider"
             ]
         }
     }
